fix for jakub
